refactor: normalize driver filenames to underscores
Converge tx-isp-*/txx-* dash names to tx_isp_*/txx_* across driver
sources, headers, includes, Kbuild, and docs. Delete tx-isp-debug.c
(isp_printf/private_* already live in tx_isp_core.c). Build the
orphaned tracer as its own tx_isp_trace.ko (separate module_init).
Add build_local.sh for a reproducible local compile baseline.

Module artifact tx-isp-t31.ko name unchanged (libimp ABI).
