+++ driver/include/tx_isp.h
@@ -21,7 +21,7 @@
 #include <linux/delay.h>
 #include <linux/miscdevice.h>
 
-#include "tx-isp-common.h"
+#include "tx_isp_common.h"
 
 #define TX_ISP_LINKFLAG_ENABLED		(0x1)
 
@@ -342,7 +342,7 @@ extern struct platform_device tx_isp_vin_platform_device;
 extern struct platform_device tx_isp_fs_platform_device;
 extern struct platform_device tx_isp_core_platform_device;
 
-/* Sensor control functions - defined in tx-isp-module.c */
+/* Sensor control functions - defined in tx_isp_module.c */
 int sensor_fps_control(int fps);
 
 #endif /* __TX_ISP_H__ */
