+++ driver/tx_isp_ae_zone.c
@@ -3,7 +3,7 @@
 #include <linux/uaccess.h>
 #include "include/tx_isp.h"
 #include "include/tx_isp_core.h"
-#include "include/tx-isp-debug.h"
+#include "include/tx_isp_debug.h"
 
 /* AE Zone Constants - Binary Ninja Reference */
 #define AE_ZONE_DATA_SIZE 0x384  /* 900 bytes - matches Binary Ninja */
