+++ driver/tx_isp_subdev_mgmt.c
@@ -10,7 +10,7 @@
 #include <linux/slab.h>
 #include "include/tx_isp.h"
 #include "include/tx_isp_vic.h"
-#include "include/tx-isp-debug.h"
+#include "include/tx_isp_debug.h"
 
 /* Subdevice type definitions */
 enum tx_isp_subdev_type {
