+++ driver/tx_isp_vic_debug.c
@@ -34,7 +34,7 @@
 #include "include/tx_isp_core.h"
 #include "include/tx_isp_vic.h"
 #include "include/tx_isp_csi.h"
-#include "include/tx-isp-debug.h"
+#include "include/tx_isp_debug.h"
 
 /* Global variables for VIC error tracking - EXACT Binary Ninja implementation */
 static uint32_t vic_err = 0;
